analyzer/utils.py
import os
from git import Repo


def analyze_repository(repo_url):
    local_path = clone_repository(repo_url)
    recommendations = "Your recommendations here."
    return recommendations

def clone_repository(repo_url):
    local_path = repo_url.split("/")[-1]
    if not os.path.exists(local_path):
        Repo.clone_from(repo_url, local_path)
    return local_path
