requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
multidict==6.1.0
openai==0.27.0
propcache==0.2.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.2.0
smmap==5.0.1
sqlparse==0.5.1
tqdm==4.66.5
urllib3==2.2.3
yarl==1.16.0
